% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okapi.R
\name{get_okapi_news}
\alias{get_okapi_news}
\title{The package for Okapi News Tracker}
\usage{
get_okapi_news(news, number)
}
\arguments{
\item{news}{type of of news to track for articles of Okapi}

\item{number}{is the number of days ago}
}
\value{
dashboard with the title of the article and some statistics about followers and comments
}
\description{
The package for Okapi News Tracker
}
\details{
An implementation of dashboard to follow the lifecycle of articles posted on the website of Okapi Radio i.e followers and comments
}
\author{
Jean Marie Cimula
}

