% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_authentication.R
\name{kobo_auth_password}
\alias{kobo_auth_password}
\title{Basic authentication to KoBo Toolbox with a username and password}
\usage{
kobo_auth_password(username, password = askpass::askpass())
}
\arguments{
\item{username}{KoBo Toolbox username.}

\item{password}{KoBo Toolbox password. Default is an interactive input of
password.}
}
\value{
System environment variables for KOBO_KF_USERNAME and KOBO_KF_PASSWORD
}
\description{
Basic authentication to KoBo Toolbox with a username and password
}
\examples{
kobo_auth_password(username = Sys.getenv("KOBO_USERNAME"),
                   password = Sys.getenv("KOBO_PASSWORD"))

}
