% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_list_assets}
\alias{kobo_list_assets}
\title{Retrieve list of assets available to specific user}
\usage{
kobo_list_assets(base_url = "https://kf.kobotoolbox.org", auth_mode = "token")

kobo_list_assets(base_url = "https://kf.kobotoolbox.org", auth_mode = "token")
}
\arguments{
\item{base_url}{KoBo Toolbox URL. Default is \url{https://kf.kobotoolbox.org}.}

\item{auth_mode}{Password or token? Default is token.}
}
\value{
A list of assets available to specific user.

A list of assets available to specific user.
}
\description{
Retrieve list of assets available to specific user

Retrieve list of assets available to specific user
}
\examples{
kobo_list_assets()

kobo_list_assets()

}
