% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odktools.R
\docType{package}
\name{odktools}
\alias{okapi}
\alias{okapi-package}
\title{Open Data Kit API client in R}
\description{
Open Data Kit (ODK) is an open-source suite of tools that helps in authoring,
collecting, and managing mobile data collection solutions. This package
provides an ODK API client in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rapidsurveys/okapi}
  \item Report bugs at \url{https://github.com/rapidsurveys/okapi/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
