% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okapi.R
\docType{package}
\name{okapi}
\alias{okapi}
\alias{okapi-package}
\title{Open Data Kit (ODK)-based Computer-assisted Personal Interview (CAPI) Tools}
\description{
Data collection using computer-assisted personal interviewing or CAPI tools
is now the standard approach for conducting surveys and studies. A wide
range of CAPI systems are currently being used. The ability to interface with
these systems helps in the overall data process. This package provides
interface functions to CAPI systems based on the Open Data Kit or ODK
technology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rapidsurveys/okapi}
  \item Report bugs at \url{https://github.com/rapidsurveys/okapi/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
