% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_type.R
\name{okapi_info_type}
\alias{okapi_info_type}
\title{The package for Okapi News Tracker}
\usage{
okapi_info_type()
}
\description{
The package for Okapi News Tracker
}
\details{
Type of news to track for articles of Okapi
}
\author{
Jean Marie Cimula
}

