% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{ona_auth_password}
\alias{ona_auth_password}
\title{Basic authentication to ONA with a username and password}
\usage{
ona_auth_password(username, password = askpass::askpass())
}
\arguments{
\item{username}{ONA username.}

\item{password}{ONA password. Default is an interactive input of password.}
}
\value{
System environment variables for ONA_USERNAME and ONA_PASSWORD
}
\description{
Basic authentication to ONA with a username and password
}
\examples{
ona_auth_password(username = "validtrial",
                  password = Sys.getenv("ONA_PASSWORD"))

}
