% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{ona_auth_token}
\alias{ona_auth_token}
\title{Basic authentication to ONA with API authentication token}
\usage{
ona_auth_token(token)
}
\arguments{
\item{token}{ONA API authentication token.}
}
\value{
System environment variables for ONA_TOKEN
}
\description{
Basic authentication to ONA with API authentication token
}
\examples{
ona_auth_token(token = Sys.getenv("ONA_TOKEN"))

}
