% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_misc.R
\name{ona_configure}
\alias{ona_configure}
\title{Configure HTTP headers}
\usage{
ona_configure(auth_mode = c("token", "password"))
}
\arguments{
\item{auth_mode}{Password or token? Default is token.}
}
\value{
An object of class request for HTTP requests
}
\description{
Configure HTTP headers
}
