% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_data.R
\name{ona_data_get}
\alias{ona_data_get}
\title{Retrieve specific dataset by form ID from specific user}
\usage{
ona_data_get(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  form_id
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{form_id}{Form identifier.}
}
\value{
A tibble of retrieved dataset.
}
\description{
Retrieve specific dataset by form ID from specific user
}
\examples{
ona_data_get(form_id = 276175)

}
