% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_form.R
\name{ona_delete_form}
\alias{ona_delete_form}
\title{Delete published form}
\usage{
ona_delete_form(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  form_id
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{form_id}{Form identifier.}
}
\value{
Delete specified form from ONA.
}
\description{
Delete published form
}
\examples{
ona_delete_form(form_id = 647324)

}
