% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_form.R
\name{ona_form_delete}
\alias{ona_form_delete}
\title{Delete published form}
\usage{
ona_form_delete(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  form_id = NULL
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{form_id}{Form identifier.}
}
\value{
Delete specified form from ONA.
}
\description{
Delete published form
}
\examples{
form_list <- ona_form_list()
form_id <- form_list[["formid"]][form_list[["title"]] == "Appearance Widgets"]
ona_form_delete(form_id = form_id)

}
