% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_form.R
\name{ona_form_list}
\alias{ona_form_list}
\title{List published forms}
\usage{
ona_form_list(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password")
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}
}
\value{
A tibble of published forms.
}
\description{
List published forms
}
\examples{
ona_form_list()

}
