% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona.R
\name{ona_list_data}
\alias{ona_list_data}
\title{Retrieve list of datasets available to specific user}
\usage{
ona_list_data(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password")
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}
}
\value{
A tibble of datasets available to specific user.
}
\description{
Retrieve list of datasets available to specific user
}
\examples{
ona_list_data()

}
