% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_projects.R
\name{ona_list_projects}
\alias{ona_list_projects}
\title{List projects}
\usage{
ona_list_projects(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password")
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}
}
\value{
A tibble of projects.
}
\description{
List projects
}
\examples{
ona_list_projects()

}
