% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_org.R
\name{ona_org_delete}
\alias{ona_org_delete}
\title{Delete registered organisation}
\usage{
ona_org_delete(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  org
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{org}{Organisation short name or username.}
}
\value{
Delete specified organisation from specific ONA account.
}
\description{
Delete registered organisation
}
\examples{
org_list <- ona_org_list()
org_name <- org_list[["org"]]
ona_org_delete(org = org_name)

}
