% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_org.R
\name{ona_org_list}
\alias{ona_org_list}
\title{List organisations}
\usage{
ona_org_list(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password")
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}
}
\value{
A tibble of organisations.
}
\description{
List organisations
}
\examples{
ona_org_list()

}
