% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_org.R
\name{ona_org_register}
\alias{ona_org_register}
\title{Register an organisation}
\usage{
ona_org_register(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  org,
  name = NULL
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{org}{Organisation short name or username.}

\item{name}{Organisation full name. If NULL (default), short name is used.}
}
\value{
Invisible. Organisation registered and created in ONA account
}
\description{
Register an organisation
}
\examples{
ona_org_register(org = "okapi_organisation")

}
