% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_project.R
\name{ona_project_delete}
\alias{ona_project_delete}
\title{Delete registered project}
\usage{
ona_project_delete(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  project_id = NULL
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{project_id}{Project identifier.}
}
\value{
Delete specified project from specific ONA account.
}
\description{
Delete registered project
}
\examples{
project_list <- ona_project_list()
project_id <- project_list[["projectid"]]
ona_project_delete(project_id = project_id)

}
