% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_projects.R
\name{ona_project_info}
\alias{ona_project_info}
\title{Get project info}
\usage{
ona_project_info(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  project_id = NULL
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{project_id}{Project identifier.}
}
\value{
A tibble of projects.
}
\description{
Get project info
}
\examples{
ona_project_info(project_id = 179619)

}
