% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_project.R
\name{ona_project_register}
\alias{ona_project_register}
\title{Register a project}
\usage{
ona_project_register(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  owner = Sys.getenv("ONA_USERNAME"),
  name = NULL,
  public = TRUE
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{owner}{Username or organisation name creating the project. Default to
username of currently authenticated account
(via `Sys.getenv(ONA_USERNAME)`). If organisation name, should be
an organisation to which currently authenticated account is the owner or
admin user.}

\item{name}{Name of the project. If NULL (default), project is given a
default name similar to that created by ONA for forms published without
a project}

\item{public}{Logical. Should the project be public? Default to TRUE.}
}
\value{
Invisible. Project registered and created in ONA account
}
\description{
Register a project
}
\examples{
ona_project_register()

}
