% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_project.R
\name{ona_project_share}
\alias{ona_project_share}
\title{Share projects with user(s)}
\usage{
ona_project_share(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  project_id = NULL,
  username = NULL,
  role = c("readonly", "dataentry", "editor", "manager"),
  email = NULL
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{project_id}{Project identifier.}

\item{username}{A character value or vector of username/s of user/s to share
a form with.}

\item{role}{A character value for the role the user/s will have on the
project. This can be \emph{readonly}, \emph{dataentry}, \emph{editor}, or \emph{manager}.}

\item{email}{Email message to send to user/s to whom project has been
shared with. If NULL (default), user/s will not be notified.}
}
\value{
Invisible. Project shared with specified users with specified
roles. A tibble of project information.
}
\description{
Share projects with user(s)
}
\examples{
ona_project_share(
  project_id = 12345,
  username = "validtrial",
  role = "readonly"
 )

}
