% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_form.R
\name{ona_publish_form}
\alias{ona_publish_form}
\title{Publish XLSForm}
\usage{
ona_publish_form(
  base_url = "https://api.ona.io",
  auth_mode = c("token", "password"),
  xls_file = NULL,
  xls_url = NULL,
  dropbox_xls_url = NULL
)
}
\arguments{
\item{base_url}{ONA URL. Default is \url{https://api.ona.io}.}

\item{auth_mode}{Password or token? Default is token.}

\item{xls_file}{Path to the XLSForm file.}

\item{xls_url}{URL to the XLSForm file.}

\item{dropbox_xls_url}{Dropbox URL to the XLSForm file.}
}
\value{
A published form on ONA.
}
\description{
Publish XLSForm
}
\examples{
ona_publish_form(
  xls_file = system.file(
    "appearance_widgets.xlsx", package = "okapi"
  )
)

}
