% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_authentication.R
\name{kobo_auth_token}
\alias{kobo_auth_token}
\title{Basic authentication to KoBo Toolbox with API authentication token}
\usage{
kobo_auth_token(token)
}
\arguments{
\item{token}{KoBo Toolbox API authentication token.}
}
\value{
System environment variables for KOBO_KF_TOKEN
}
\description{
Basic authentication to KoBo Toolbox with API authentication token
}
\examples{
kobo_auth_token(token = Sys.getenv("KOBO_TOKEN"))

}
